+++ moshi/moshi/lang.py
@@ -6,8 +6,8 @@ from enum import Enum
 import pyttsx3
 from loguru import logger
 
-from moshimoshi import think
-from moshimoshi.base import Message, Role
+from moshi import think
+from moshi.base import Message, Role
 
 N_COMPLETIONS = os.getenv("MOSHI_LANGUAGE_DETECT_COMPLETIONS", 3)