+++ moshi/tests/test_listen.py
@@ -3,14 +3,14 @@ from unittest import mock
 import pytest
 from openai.error import RateLimitError
 
-from moshimoshi import listen
+from moshi import listen
 
 @pytest.mark.usefixtures("mock_rec_listen")
 def test_get_audio_from_mic(audio):
     raudio = listen._get_audio_from_mic()
     assert audio == raudio
 
-@mock.patch('moshimoshi.listen.rec.recognize_whisper_api', lambda _: "test")
+@mock.patch('moshi.listen.rec.recognize_whisper_api', lambda _: "test")
 def test_transcribe_audio(audio):
     transcription = listen._transcribe_audio(audio)
     assert transcription == "test"
@@ -27,7 +27,7 @@ def test_dialogue_from_mic():
 
 @pytest.mark.openai
 @pytest.mark.usefixtures("mock_rec_listen")
-@mock.patch('moshimoshi.listen.rec.recognize_whisper_api', lambda _: "test")
+@mock.patch('moshi.listen.rec.recognize_whisper_api', lambda _: "test")
 @pytest.mark.xfail(reason="Rate limits from OpenAI", raises=RateLimitError)
 def test_dialogue_from_mic_openai():
     assert listen.dialogue_from_mic() == "test"